// convex/_generated/api.d.ts
declare const fullApi: ApiFromModules<{
  alerts: typeof alerts;
  auth: typeof auth;
  cameras: typeof cameras;
  crons: typeof crons;
  email: typeof email;
  http: typeof http;
  search: typeof search;
  videos: typeof videos;
}>;

/**
 * A utility for referencing Convex functions in your app's public API.
 *
 * Usage:
 * ```js
 * const myFunctionReference = api.myModule.myFunction;
 * ```
 */
export declare const api: FilterApi<
  typeof fullApi,
  FunctionReference<any, "public">
>;